+++ src/xcc/aot/slice.py
@@ -49,6 +49,7 @@ _NATIVE_EMITTED_LEAF_FUNCTIONS = {
     "xcc.cc_driver._aot_exec_argv",
     "xcc.cc_driver._aot_read_text_file",
     "xcc.cc_driver._aot_write_text_file",
+    "xcc.codegen._llvm_print_module_to_string",
     "xcc.lexer._aot_error_summary_for_source",
     "xcc.lexer._aot_header_summary_for_source",
     "xcc.lexer._aot_token_summary_for_source",